Assignment 11
I started research on my topic and found 5 useful articles that I will use to further progress my analysis. I also started writing my report. I finished the Introduction and Data sections of the report. Later this week I will work on the methods and results sections of the paper seeing that this will be where the bulk of the paper will lie.